+++ src/components/BlogCreate.tsx
@@ -1,14 +1,14 @@
-import axios, { AxiosResponse } from 'axios';
-import { IBlogCreateDto, IBlogEntity } from 'blog-common-1.0';
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router';
-import { getJwt } from '../helpers/helper';
+import { IBlogCreateDto } from "blog-common-1.0";
+import React, { useState } from "react";
+import { useNavigate } from "react-router";
+import { createBlogApiCallFunctionU } from "../api functions/blogs.api.calls.function";
+import { getJwt } from "../helpers/helper";
 
 export default function BlogCreate() {
   const [formData, setFormData] = useState<IBlogCreateDto>({
-    title: '',
-    content: '',
-    keywords: '',
+    title: "",
+    content: "",
+    keywords: "",
   });
   const [newBlog, setNewBlog] = useState<IBlogCreateDto | null>(null);
 
@@ -26,40 +26,41 @@ export default function BlogCreate() {
   };
 
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    console.log('this is the blog data', formData);
+    await createBlogApiCallFunctionU(e, formData, setNewBlog);
 
-    const token: string | null = getJwt();
+    // e.preventDefault();
+    // console.log('this is the blog data', formData);
 
-    try {
-      const response: AxiosResponse<IBlogEntity> = await axios.post(
-        'http://localhost:3000/api/blog',
-        formData, {
-        headers: {
-          Authorization: `Bearer ${token}`, // Add token to header
-        }
-      }
-      );
-      console.log('this is the response', response);
-      const { data } = response;
-      console.log('this is the new blog', data);
-      setNewBlog(data);
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        // console.log("this is the response", response)
-        console.error(error.response?.data); // Log the backend error response
-      } else {
-        console.error('Unexpected error:', error);
-      }
-    }
+    // const token: string | null = getJwt();
+
+    // try {
+    //   const response: AxiosResponse<IBlogEntity> = await axios.post(
+    //     'http://localhost:3000/api/blog',
+    //     formData, {
+    //     headers: {
+    //       Authorization: `Bearer ${token}`, // Add token to header
+    //     }
+    //   }
+    //   );
+    //   console.log('this is the response', response);
+    //   const { data } = response;
+    //   console.log('this is the new blog', data);
+    //   setNewBlog(data);
+    // } catch (error) {
+    //   if (axios.isAxiosError(error)) {
+    //     // console.log("this is the response", response)
+    //     console.error(error.response?.data); // Log the backend error response
+    //   } else {
+    //     console.error('Unexpected error:', error);
+    //   }
+    // }
   };
 
   const routeToUpdateBlog = () => {
-    console.log("click event for routing to update blog page")
-    const navigate = useNavigate()
-    navigate("/api/updateBlog")
-
-  }
+    console.log("click event for routing to update blog page");
+    const navigate = useNavigate();
+    navigate("/api/updateBlog");
+  };
 
   return (
     <>
@@ -100,7 +101,7 @@ export default function BlogCreate() {
 
       <button onClick={routeToUpdateBlog}>
         {/* <Link to={"/api/updateBlog"}> */}
-          Update Blog
+        Update Blog
         {/* </Link> */}
       </button>
